requirements.txt
torch
numpy
opencv-python
pillow
matplotlib
tqdm
torchvision
einops
python-dotenv
scikit-image